start mdx, fix story names, fix compact padding
